Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,22 +1,22 @@
 #include <stdio.h>
 int main() {
-    float a,b;
+    int a,b;
     char c;
-   scanf("%f %f %c",&a,&b,&c);
+   scanf("%d %d %c",&a,&b,&c);
 
    switch(c){
     case '+':
-    printf("%.0f",a+b);
+    printf("%d",a+b);
     break;
      case '-':
-    printf("%.0f",a-b);
+    printf("%d",a-b);
     break;
      case '*':
-    printf("%.0f",a*b);
+    printf("%d",a*b);
     break;
      case '/':
      if(b>0){
-    printf("%.0f",a/b);
+    printf("%d",a/b);
     break;
      }
      else{
